Move monkeypatch back to onnx
